Define default layers for individual UpSet components

Define and modify default layers for individual UpSet components

default_set_layers(
  ...,
  fill = NULL,
  f = comma,
  expand = 0.2,
  hjust = 1.1,
  size = 3.5,
  name = "Set Size",
  dry_run = FALSE
)

default_intersect_layers(
  ...,
  fill = NULL,
  f = comma,
  expand = 0.05,
  vjust = -0.5,
  size = 3.5,
  name = "Intersection Size",
  dry_run = FALSE
)

default_grid_layers(
  ...,
  colour = NULL,
  fill = NULL,
  light = "grey80",
  dark = "grey23",
  shape = 19,
  size = 4,
  name = NULL,
  dry_run = FALSE
)

Arguments

...

additional layers to include alongside default layers. Will be added after the default layers

fill

Column to fill set bars by. Can be 'set' or another column within the main data object

f

Function for labelling set sizes and the set x-axis

expand

axis expansion

hjust, vjust

Passed to respective elements for simple adjustment of either set or intersection sizes

size

Passed to labels for sets and intersections, and to point sizes for the intersections matrix

name

Main axis title

dry_run

Set as TRUE to view the unevaluated layers which are defined as the defaults. Additional layers passed through the ellipsis will not be included as part of a dry_run

colour

Primarily used for highlighting points and segments in the intersections matrix

light, dark

default colours for empty intersections (light) and for both non-empty intersections and segments (dark)

shape

Point shape passed to the intersections matrix

Value

List of ggplot2 elements

Details

These functions define the default layers for inclusion on UpSet plots.

The returned object is a list with a series of layers, scales, themes etc which represent the default plotting layers for each of the sets, intersections and intersections matrix (grid) panels.

A series of common arguments have been defined to enable common modifications without recreating the list from scratch. These include modifying the mapping to fill, axis expansion to better accommodate labels, labelling functions for set/intersection sizes, and axis titles.

Additional layers, such as scale_fill_* elements, guides or themes, can be simply included by passing to the function, without any requirement for naming, and are handled by the ellipsis.

The entire command used to create default layers can be shown by calling each function using the argument dry_run = TRUE. This can be helpful for creating custom layers, by starting with then modifying the defaults.

The returned object is a simple list, and are easily modifiable using simple list operations.

Examples

# View the un-evaluated list of default layers for the sets
default_set_layers(dry_run = TRUE)
#> list(aes(y = set), geom_bar(bar_aes), geom_text(aes(x = n, label = f(n)), 
#>     hjust = hjust, size = size), scale_x_reverse(expand = c(expand, 
#>     0, 0, 0), name = name, labels = f), scale_y_discrete(position = "right", 
#>     name = NULL, labels = NULL), theme(axis.text.y.right = element_text(hjust = 0.5), 
#>     axis.ticks.y.right = element_blank(), margins = margin(5.5, 
#>         0, 5.5, 5.5)))

# Create set layers colouring by set name, and hiding the legend
set_list <- default_set_layers(
  fill = "set", scale_fill_brewer(palette = "Set1"), guides(fill = guide_none())
)
sapply(set_list, is)
#> [1] "ggplot2::mapping"        "LayerInstance"          
#> [3] "LayerInstance"           "ScaleContinuousPosition"
#> [5] "ScaleDiscretePosition"   "theme"                  
#> [7] "ScaleDiscrete"           "Guides"